+++ CHANGELOG.md
@@ -7,11 +7,12 @@ Master
 ------
 
 * BC BREAK: Refactored ExclusionStrategy interface
+* BC BREAK: Setting a cache directory no longer enables caching, use enableCache() method on builder
+* BC BREAK: Removed ArrayList/HashMap type adapters
+* BC BREAK: Removed support for 'JsonElement' type, must use full class name
 * FIXED: Infinite loop on circular references when excluding property
 * ADDED: Type Adapter caching
 * ADDED: Support for subclasses of DateTime
-* REMOVED: ArrayList/HashMap type adapters
-* REMOVED: Support for 'JsonElement' type, must use full class name
 
 v0.1.0
 ------